# docs/openapi.json
{
  "components": {
    "schemas": {
...
      "ErrorResponse": {
        "description": "Fehlerantwort (``HTTPException``): ``detail`` ist die Meldung.",
        "properties": {
          "detail": {
            "title": "Detail",
            "type": "string"
          }
        },
        "required": [
          "detail"
        ],
        "title": "ErrorResponse",
        "type": "object"
      },
...
      "RunCreated": {
        "properties": {
          "run_id": {
            "title": "Run Id",
            "type": "string"
          },
          "status": {
            "title": "Status",
            "type": "string"
          }
        },
        "required": [
          "run_id",
          "status"
        ],
        "title": "RunCreated",
        "type": "object"
      },
...
      "RunHistory": {
        "description": "FA89: die Läufe des Nutzers, die das Backend noch hält (neuester zuerst).",
        "properties": {
          "max_retained": {
            "title": "Max Retained",
            "type": "integer"
          },
          "runs": {
            "items": {
              "$ref": "#/components/schemas/RunSummary"
            },
            "title": "Runs",
            "type": "array"
          }
        },
        "required": [
          "max_retained"
        ],
        "title": "RunHistory",
        "type": "object"
      },
...
    },
    "securitySchemes": {
      "HTTPBearer": {
        "description": "Token aus POST /api/auth/login als 'Authorization: Bearer <token>'. Nur wirksam, wenn GEOFACT_WEB_USERS gesetzt ist (FA24); ohne konfigurierte Nutzer wird kein Token geprüft.",
        "scheme": "bearer",
        "type": "http"
      },
      "QueryToken": {
        "description": "Dasselbe Token als Query-Parameter - nur für Ereignisstrom, Rasterbild und Download, die der Browser ohne Authorization-Header abruft. Nur wirksam, wenn GEOFACT_WEB_USERS gesetzt ist (FA24); ohne konfigurierte Nutzer wird kein Token geprüft.",
        "in": "query",
        "name": "token",
        "type": "apiKey"
      }
    }
  },
  "info": {
    "description": "Backend-API zur interaktiven Demonstration des GeoFACT-Frameworks.",
    "title": "GeoFACT Web API",
    "version": "0.1.0"
  },
  "openapi": "3.1.0",
  "paths": {
...
    "/api/runs": {
      "get": {
        "description": "Die Läufe des Nutzers, die das Backend noch hält, neuester zuerst. Läufe liegen nur im Arbeitsspeicher: nach einem Neustart ist die Liste leer, und von den abgeschlossenen bleiben höchstens max_retained.",
        "operationId": "list_runs_api_runs_get",
        "responses": {
          "200": {
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/RunHistory"
                }
              }
            },
            "description": "Successful Response"
          },
          "401": {
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/ErrorResponse"
                }
              }
            },
            "description": "Token fehlt, ist ungültig oder abgelaufen. Nur wirksam, wenn GEOFACT_WEB_USERS gesetzt ist (FA24); ohne konfigurierte Nutzer wird kein Token geprüft."
          }
        },
        "security": [
          {
            "HTTPBearer": []
          }
        ],
        "summary": "Läufe auflisten",
        "tags": [
          "runs"
        ]
      },
      "post": {
        "description": "Prüft die Konfiguration und startet den Lauf im Hintergrund. Fortschritt über GET /api/runs/{run_id} oder den Ereignisstrom.",
        "operationId": "create_run_api_runs_post",
        "requestBody": {
          "content": {
            "application/json": {
              "schema": {
                "$ref": "#/components/schemas/RunRequest"
              }
            }
          },
          "required": true
        },
        "responses": {
          "200": {
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/RunCreated"
                }
              }
            },
            "description": "Successful Response"
          },
          "401": {
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/ErrorResponse"
                }
              }
            },
            "description": "Token fehlt, ist ungültig oder abgelaufen. Nur wirksam, wenn GEOFACT_WEB_USERS gesetzt ist (FA24); ohne konfigurierte Nutzer wird kein Token geprüft."
          },
          "422": {
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/RejectedRequestResponse"
                }
              }
            },
            "description": "Konfiguration ungültig (detail.issues) oder Anfrage formal ungültig."
          },
          "429": {
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/ErrorResponse"
                }
              }
            },
            "description": "Zu viele gleichzeitige Läufe dieses Nutzers."
          }
        },
        "security": [
          {
            "HTTPBearer": []
          }
        ],
        "summary": "Lauf starten",
        "tags": [
          "runs"
        ]
      }
    },
    "/api/runs/{run_id}": {
      "get": {
        "description": "Zustand des Laufs und jedes Schritts, mit Warnungen, Ladezeiten, Region, CRS-Herkunft der Layer und Lizenzen.",
        "operationId": "run_status_api_runs__run_id__get",
        "parameters": [
          {
            "in": "path",
            "name": "run_id",
            "required": true,
            "schema": {
              "title": "Run Id",
              "type": "string"
            }
          }
        ],
        "responses": {
          "200": {
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/RunStatus"
                }
              }
            },
            "description": "Successful Response"
          },
          "401": {
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/ErrorResponse"
                }
              }
            },
            "description": "Token fehlt, ist ungültig oder abgelaufen. Nur wirksam, wenn GEOFACT_WEB_USERS gesetzt ist (FA24); ohne konfigurierte Nutzer wird kein Token geprüft."
          },
          "404": {
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/ErrorResponse"
                }
              }
            },
            "description": "Lauf unbekannt oder gehört einem anderen Nutzer. Läufe liegen nur im Arbeitsspeicher; nach einem Neustart des Backends sind sie weg."
          },
          "422": {
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/HTTPValidationError"
                }
              }
            },
            "description": "Validation Error"
          }
        },
        "security": [
          {
            "HTTPBearer": []
          }
        ],
        "summary": "Lauf-Status",
        "tags": [
          "runs"
        ]
      }
    },
    "/api/runs/{run_id}/cancel": {
      "post": {
        "description": "Der Lauf gilt sofort als abgebrochen (Status cancelled, Ereignis run_cancelled, Ende des Ereignisstroms). Ein gerade laufender Baustein endet im Hintergrund an der nächsten Schrittgrenze; sein Ergebnis wird verworfen. cancelling=false: der Lauf war schon beendet.",
        "operationId": "cancel_run_api_runs__run_id__cancel_post",
        "parameters": [
          {
            "in": "path",
            "name": "run_id",
            "required": true,
            "schema": {
              "title": "Run Id",
              "type": "string"
            }
          }
        ],
        "responses": {
          "200": {
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/RunCancelResponse"
                }
              }
            },
            "description": "Successful Response"
          },
          "401": {
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/ErrorResponse"
                }
              }
            },
            "description": "Token fehlt, ist ungültig oder abgelaufen. Nur wirksam, wenn GEOFACT_WEB_USERS gesetzt ist (FA24); ohne konfigurierte Nutzer wird kein Token geprüft."
          },
          "404": {
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/ErrorResponse"
                }
              }
            },
            "description": "Lauf unbekannt oder gehört einem anderen Nutzer. Läufe liegen nur im Arbeitsspeicher; nach einem Neustart des Backends sind sie weg."
          },
          "422": {
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/HTTPValidationError"
                }
              }
            },
            "description": "Validation Error"
          }
        },
        "security": [
          {
            "HTTPBearer": []
          }
        ],
        "summary": "Lauf abbrechen",
        "tags": [
          "runs"
        ]
      }
    },
    "/api/runs/{run_id}/download": {
      "get": {
        "description": "ZIP mit den geschriebenen Ausgaben und scenario.yaml, dazu ATTRIBUTION.txt, parameters.yaml und SKIPPED_OUTPUTS.txt, soweit vorhanden.",
        "operationId": "download_api_runs__run_id__download_get",
        "parameters": [
          {
            "in": "path",
            "name": "run_id",
            "required": true,
            "schema": {
              "title": "Run Id",
              "type": "string"
            }
          }
        ],
        "responses": {
          "200": {
            "content": {
              "application/zip": {
                "schema": {
                  "type": "string"
                }
              }
            },
            "description": "ZIP-Archiv."
          },
          "401": {
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/ErrorResponse"
                }
              }
            },
            "description": "Token fehlt, ist ungültig oder abgelaufen. Nur wirksam, wenn GEOFACT_WEB_USERS gesetzt ist (FA24); ohne konfigurierte Nutzer wird kein Token geprüft."
          },
          "404": {
            "content": {
              "application/zip": {
                "schema": {
                  "$ref": "#/components/schemas/ErrorResponse"
                }
              }
            },
            "description": "Lauf unbekannt oder gehört einem anderen Nutzer. Läufe liegen nur im Arbeitsspeicher; nach einem Neustart des Backends sind sie weg."
          },
          "409": {
            "content": {
              "application/zip": {
                "schema": {
                  "$ref": "#/components/schemas/ErrorResponse"
                }
              }
            },
            "description": "Lauf ist noch nicht abgeschlossen."
          },
          "422": {
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/HTTPValidationError"
                }
              }
            },
            "description": "Validation Error"
          }
        },
        "security": [
          {
            "HTTPBearer": []
          },
          {
            "QueryToken": []
          }
        ],
        "summary": "Ergebnis herunterladen",
        "tags": [
          "runs"
        ]
      }
    },
    "/api/runs/{run_id}/events": {
      "get": {
        "description": "Server-Sent Events: zuerst status (Lauf-Status vor dem ersten Ereignis), dann je Ereignis des Laufs ein progress, zum Schluss status (Endstand) und end.",
        "operationId": "run_events_api_runs__run_id__events_get",
        "parameters": [
          {
            "in": "path",
            "name": "run_id",
            "required": true,
            "schema": {
              "title": "Run Id",
              "type": "string"
            }
          }
        ],
        "responses": {
          "200": {
            "content": {
              "text/event-stream": {
                "schema": {
                  "type": "string"
                }
              }
            },
            "description": "Ereignisstrom bis end."
          },
          "401": {
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/ErrorResponse"
                }
              }
            },
            "description": "Token fehlt, ist ungültig oder abgelaufen. Nur wirksam, wenn GEOFACT_WEB_USERS gesetzt ist (FA24); ohne konfigurierte Nutzer wird kein Token geprüft."
          },
          "404": {
            "content": {
              "text/event-stream": {
                "schema": {
                  "$ref": "#/components/schemas/ErrorResponse"
                }
              }
            },
            "description": "Lauf unbekannt oder gehört einem anderen Nutzer. Läufe liegen nur im Arbeitsspeicher; nach einem Neustart des Backends sind sie weg."
          },
          "422": {
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/HTTPValidationError"
                }
              }
            },
            "description": "Validation Error"
          }
        },
        "security": [
          {
            "HTTPBearer": []
          },
          {
            "QueryToken": []
          }
        ],
        "summary": "Live-Fortschritt (Ereignisstrom)",
        "tags": [
          "runs"
        ]
      }
    },
...
  },
...
}